requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.2.0
django-sortedm2m==3.1.1
django-autoslug==1.9.9
django-jazzmin==2.6.0
djoser==2.2.2
django-debug-toolbar==4.2.0
django-filter==23.5
argon2-cffi==23.1.0
